Cleanup -- plus added DDOS hack 5b5b5b5 handler code

+++ src/libs/include/sfstring.h
@@ -611,7 +611,7 @@ namespace qblocks {
 
     //----------------------------------------------------------------------------
     inline SFString shorten(const SFString& in, size_t x) {
-        return padRight(in.length()>x-3 ? in.Left(x-3) + "..." : in, x);
+        return padRight(in.length()>x-3 ? in.Left(x-3) + "..." : in, (uint32_t)x);
     }
 
     //--------------------------------------------------------------------
@@ -641,7 +641,7 @@ namespace qblocks {
     inline SFString StripAny(const SFString& str, const SFString& any) {
         SFString ret = str;
         while (endsWithAny(ret, any) || startsWithAny(ret, any)) {
-            for (size_t i = 0 ; i < any.length() ; i++)
+            for (uint32_t i = 0 ; i < any.length() ; i++)
                 ret = Strip(ret, any[i]);
         }
         return ret;
